backend/api/serializers/employee.py
from rest_framework import serializers

from api.views.employee import Employee

class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = ['id', 'name', 'email', 'is_available', 'company']

